+++ packages/cli/api-importers/openapi-to-ir/src/3.1/paths/operations/OperationConverter.ts
@@ -595,7 +595,8 @@ export class OperationConverter extends AbstractOperationConverter {
                 availability: undefined,
                 docs: undefined,
                 env: undefined,
-                v2Examples: undefined
+                v2Examples: undefined,
+                clientDefault: undefined
             };
 
             switch (authScheme.type) {