+++ packages/cli/api-importers/v3-importer-commons/src/converters/abstract/AbstractParameterConverter.ts
@@ -84,7 +84,8 @@ export abstract class AbstractParameterConverter<
                             schema: parameterSchemaWithExampleOverride ?? schema
                         }),
                         availability,
-                        explode: this.getExplodeForQueryParameter()
+                        explode: this.getExplodeForQueryParameter(),
+                        clientDefault: undefined
                     },
                     inlinedTypes
                 };
@@ -102,7 +103,8 @@ export abstract class AbstractParameterConverter<
                         v2Examples: this.convertParameterExamples({
                             schema: parameterSchemaWithExampleOverride ?? schema
                         }),
-                        availability
+                        availability,
+                        clientDefault: undefined
                     },
                     inlinedTypes
                 };
@@ -118,7 +120,8 @@ export abstract class AbstractParameterConverter<
                         v2Examples: this.convertParameterExamples({
                             schema: parameterSchemaWithExampleOverride ?? schema
                         }),
-                        explode: this.getExplodeForPathParameter()
+                        explode: this.getExplodeForPathParameter(),
+                        clientDefault: undefined
                     },
                     inlinedTypes
                 };