+++ packages/ir-sdk/src/sdk/serialization/resources/http/types/PathParameter.ts
@@ -6,6 +6,7 @@ import * as serializers from "../../../index.js";
 import { Name } from "../../commons/types/Name.js";
 import { WithDocs } from "../../commons/types/WithDocs.js";
 import { V2SchemaExamples } from "../../examples/types/V2SchemaExamples.js";
+import { Literal } from "../../types/types/Literal.js";
 import { VariableId } from "../../variables/types/VariableId.js";
 import { PathParameterLocation } from "./PathParameterLocation.js";
 
@@ -16,6 +17,7 @@ export const PathParameter: core.serialization.ObjectSchema<serializers.PathPara
             valueType: core.serialization.lazy(() => serializers.TypeReference),
             location: PathParameterLocation,
             variable: VariableId.optional(),
+            clientDefault: Literal.optional(),
             v2Examples: V2SchemaExamples.optional(),
             explode: core.serialization.boolean().optional(),
         })
@@ -27,6 +29,7 @@ export declare namespace PathParameter {
         valueType: serializers.TypeReference.Raw;
         location: PathParameterLocation.Raw;
         variable?: VariableId.Raw | null;
+        clientDefault?: Literal.Raw | null;
         v2Examples?: V2SchemaExamples.Raw | null;
         explode?: boolean | null;
     }