+++ seed/rust-sdk/bytes-upload/src/lib.rs
@@ -13,10 +13,7 @@
 //!         ..Default::default()
 //!     };
 //!     let client = BytesUploadClient::new(config).expect("Failed to build client");
-//!     client
-//!         .service
-//!         .upload(&todo!("Invalid bytes value"), None)
-//!         .await;
+//!     client.service.upload(&vec![], None).await;
 //! }
 //! ```
 //!