+++ seed/swift-sdk/x-fern-default/Sources/Schemas/TestGetResponse.swift
@@ -1,0 +1,32 @@
+import Foundation
+
+public struct TestGetResponse: Codable, Hashable, Sendable {
+    public let message: String?
+    /// Additional properties that are not explicitly defined in the schema
+    public let additionalProperties: [String: JSONValue]
+
+    public init(
+        message: String? = nil,
+        additionalProperties: [String: JSONValue] = .init()
+    ) {
+        self.message = message
+        self.additionalProperties = additionalProperties
+    }
+
+    public init(from decoder: Decoder) throws {
+        let container = try decoder.container(keyedBy: CodingKeys.self)
+        self.message = try container.decodeIfPresent(String.self, forKey: .message)
+        self.additionalProperties = try decoder.decodeAdditionalProperties(using: CodingKeys.self)
+    }
+
+    public func encode(to encoder: Encoder) throws -> Void {
+        var container = encoder.container(keyedBy: CodingKeys.self)
+        try encoder.encodeAdditionalProperties(self.additionalProperties)
+        try container.encodeIfPresent(self.message, forKey: .message)
+    }
+
+    /// Keys for encoding/decoding struct properties.
+    enum CodingKeys: String, CodingKey, CaseIterable {
+        case message
+    }
+}